+++ R/spatial_power.R
@@ -21,6 +21,7 @@
 #' @param l_control Optional. A single positive number, a vector of positive numbers, a function(x,y, ...), or a pixel image. Intensity of the Poisson process for control clusters. Ignored if \code{samp_control="uniform"}, \code{samp_control="systematic"}, \code{samp_control="MVN"}, or \code{samp_control="CSR"}.
 #' @param e_control Optional. A single non-negative number for the size of the expansion of the simulation window for generating parent points. Ignored if \code{samp_control!="clustered"}.
 #' @param alpha Optional. Numeric value of the critical p-value (default=0.05).
+#' @param p_correct Optional. Character string specifying whether to apply a correction for multiple comparisons including a False Discovery Rate \code{p_correct = "FDR"}, a Sidak correction \code{p_correct = "uncorrelated Sidak"}, and a Bonferroni correction \code{p_correct = "uncorrelated Bonferroni"}. If \code{p_correct = "none"} (the default), then no correction is applied. 
 #' @param parallel Logical. If TRUE, will execute the function in parallel. If FALSE (the default), will not execute the function in parallel.
 #' @param n_core Optional. Integer specifying the number of CPU cores on current host to use for parallelization (the default is 2 cores).
 #' @param verbose Logical. If TRUE (the default), will print function progress during execution. If FALSE, will not print.
@@ -52,6 +53,8 @@
 #' 
 #' The function computes a one-sided hypothesis test for case clustering (\code{alpha = 0.05} by default). The function also computes a two-sided hypothesis test for case clustering and control clustering (lower tail = 0.025 and upper tail = 0.975).
 #' 
+#' The function has functionality for a correction for multiple testing. If \code{p_correct = "FDR"}, calculates a False Discovery Rate by Benjamini and Hochberg. If \code{p_correct = "Sidak"}, calculates a Sidak correction. If \code{p_correct = "Bonferroni"}, calculates a Bonferroni correction. If \code{p_correct = "none"} (the default), then the function does not account for multiple testing and uses the uncorrected \code{alpha} level. See the internal \code{pval_correct} function documentation for more details.
+#' 
 #' @return An object of class "list". This is a named list with the following components:
 #' 
 #' \describe{
@@ -69,6 +72,7 @@
 #' \item{\code{bandw}}{Vector of length \code{sim_total} of the bandwidth (of numerator) used in each iteration.}
 #' \item{\code{s_obs}}{Vector of length \code{sim_total} of the global s statistic.}
 #' \item{\code{t_obs}}{Vector of length \code{sim_total} of the global t statistic.}
+#' \item{\code{alpha}}{Vector of length \code{sim_total} of the (un)corrected critical p-values.}
 #' }
 #' 
 
@@ -108,7 +112,8 @@ spatial_power <- function(win = spatstat.geom::unit.square(),
                           s_case = NULL, s_control = NULL,
                           l_case = NULL, l_control = NULL,
                           e_control = NULL,
-                          alpha = 0.05, 
+                          alpha = 0.05,
+                          p_correct = c("none", "FDR", "Sidak", "Bonferroni"),
                           verbose = TRUE,
                           parallel = FALSE,
                           n_core = 2,
@@ -166,6 +171,9 @@ spatial_power <- function(win = spatstat.geom::unit.square(),
      }
     s_control <- unlist(l)
   }
+  
+  match.arg(p_correct, choices = c("none", "FDR", "Sidak", "Bonferroni"))
+  if (is.null(p_correct)) { p_correct <- "none"}
   
   # marked uniform disc ppp with user-specified radius for cases
   rcluster_case <- function(x0, y0, rad, n, scalar, lamb, wind, types = "case", ...) {
@@ -285,6 +293,7 @@ spatial_power <- function(win = spatstat.geom::unit.square(),
                               .init = list(list(), list(), list(),
                                            list(), list(), list(),
                                            list(), list(), list(),
+                                           list(), list(), list(),
                                            list(), list())) %fun% {
     
     # Progress bar
@@ -344,6 +353,23 @@ spatial_power <- function(win = spatstat.geom::unit.square(),
     sim_risk <- as.vector(t(obs_lrr$rr$v))
     sim_pval <- as.vector(t(obs_lrr$P$v))
     
+    if (p_correct != "none") {
+      alpha_correct <- pval_correct(input = obs_lrr, type = p_correct, alpha = alpha)
+      
+      #### Case and Control (lower and upper tail)
+      lower_tail <- alpha_correct/2
+      upper_tail <- 1 - lower_tail
+      pval_sig_cascon <- sapply(sim_pval, function(x) ifelse(x < lower_tail | x > upper_tail,
+                                                             TRUE,
+                                                             FALSE))
+      #### Case only (lower tail only)
+      pval_sig_cas <- sapply(sim_pval, function(x) ifelse(x < alpha_correct, TRUE, FALSE))
+    } else {
+      alpha_correct <- alpha
+      pval_sig_cascon <- "Uncorrected"
+      pval_sig_cas <- "Uncorrected"
+    }
+    
     ### Estimated global test statistics
     #### Global maximum relative risk: H0 = 1
     s_obs <- max(exp(obs_lrr$rr$v[!is.na(obs_lrr$rr$v)]))
@@ -373,7 +399,10 @@ spatial_power <- function(win = spatstat.geom::unit.square(),
                         "n_con" = con$n,
                         "bandw" = obs_lrr$f$h0,
                         "s_obs" = s_obs,
-                        "t_obs" = t_obs)
+                        "t_obs" = t_obs,
+                        "alpha_correct" = alpha_correct,
+                        "pval_sig_cascon" = pval_sig_cascon,
+                        "pval_sig_cas" = pval_sig_cas)
     return(par_results)
   }
   
@@ -395,7 +424,12 @@ spatial_power <- function(win = spatstat.geom::unit.square(),
   rr_sd <- apply(sim_rr_dat, 1, sd, na.rm = TRUE) # standard deviation log relative risk
   
   ## Calculate proportion of tests were significant
-  ### Significance level is user-specified
+  ### Correction for multiple testing
+  if (p_correct != "none") {
+    pval_sig_cascon <- out_par[[13]]
+    pval_sig_cas <- out_par[[14]]
+  } else {
+  ### Uncorrected for multiple testing
   #### Case and Control (lower and upper tail)
   lower_tail <- alpha/2
   upper_tail <- 1 - lower_tail
@@ -403,11 +437,14 @@ spatial_power <- function(win = spatstat.geom::unit.square(),
                                                          TRUE,
                                                          FALSE),
                             how = "replace")
-  pval_count_cascon <- rowSums(do.call(cbind, pval_sig_cascon), na.rm = TRUE)
-  pval_prop_wNA_cascon <- sapply(pval_count_cascon, FUN = function(x, y = sim_total) (x / y))
   #### Case only (lower tail only)
   pval_sig_cas <- rapply(sim_pval, function(x) ifelse(x < alpha, TRUE, FALSE),
                          how = "replace")
+  }
+  
+  pval_count_cascon <- rowSums(do.call(cbind, pval_sig_cascon), na.rm = TRUE)
+  pval_prop_wNA_cascon <- sapply(pval_count_cascon, FUN = function(x, y = sim_total) (x / y))
+  
   pval_count_cas <- rowSums(do.call(cbind, pval_sig_cas), na.rm = TRUE)
   pval_prop_wNA_cas <- sapply(pval_count_cas, FUN = function(x, y = sim_total) (x / y))
   
@@ -435,5 +472,6 @@ spatial_power <- function(win = spatstat.geom::unit.square(),
                   "n_con" = unlist(out_par[[8]]),
                   "bandw" = unlist(out_par[[9]]),
                   "s_obs" = unlist(out_par[[10]]),
-                  "t_obs" = unlist(out_par[[11]]))
+                  "t_obs" = unlist(out_par[[11]]),
+                  "alpha" = unlist(out_par[[12]]))
 }